+++ cloudwanderer/aws_interface/__init__.py
@@ -1,5 +1,6 @@
 """The CloudWanderer AWS Interface."""
 from .interface import CloudWandererAWSInterface
 from .session import CloudWandererBoto3Session
+from .models import AWSResourceTypeFilter
 
-__all__ = ["CloudWandererAWSInterface", "CloudWandererBoto3Session"]
+__all__ = ["CloudWandererAWSInterface", "CloudWandererBoto3Session", "AWSResourceTypeFilter"]